simplify navigation and question context providers
